requirements.txt
ffmpeg-python
google-cloud-speech